Moved passport to its own utility file (i.e. out of app.js). Moved module existence checker to a utility file.
